+++ pymobiledevice3/cli/pcap.py
@@ -1,3 +1,6 @@
+import ipaddress
+import logging
+import time
 from datetime import datetime
 from pathlib import Path
 from typing import Annotated, Optional
@@ -7,7 +10,9 @@ from pygments import formatters, highlight, lexers
 from typer_injector import InjectingTyper
 
 from pymobiledevice3.cli.cli_common import ServiceProviderDep, print_hex, user_requested_colored_output
-from pymobiledevice3.services.pcapd import PcapdService
+from pymobiledevice3.services.pcapd import CrossPlatformAddressFamily, PcapdService
+
+logger = logging.getLogger(__name__)
 
 cli = InjectingTyper(
     name="pcap",
@@ -76,3 +81,53 @@ def pcap(
 
     for packet in packets_generator:
         print_packet(packet)
+
+
+@cli.command("wifi-ip")
+def wifi_ip(
+    service_provider: ServiceProviderDep,
+    timeout: Annotated[
+        int,
+        typer.Option(help="Timeout in seconds to wait for a connection (0 = no timeout)"),
+    ] = 5,
+) -> None:
+    """
+    Get the device's WiFi IP address via USB.
+
+    Captures network packets and extracts the first private IPv4 source address,
+    which corresponds to the device's WiFi IP.
+
+    This does not require developer mode to be enabled.
+
+    \b
+    Examples:
+        pymobiledevice3 pcap wifi-ip
+        pymobiledevice3 pcap wifi-ip --timeout 10
+    """
+    ETHERNET_HEADER_LEN = 14
+    IPV4_SRC_OFFSET = 12
+
+    service = PcapdService(lockdown=service_provider)
+    start_time = time.time()
+
+    for packet in service.watch():
+        if timeout > 0 and (time.time() - start_time) > timeout:
+            logger.error(f"Timeout after {timeout} seconds. No WiFi IP found.")
+            logger.info("Tip: Generate network activity on the device (e.g., open a webpage)")
+            raise typer.Exit(code=1)
+
+        if packet.protocol_family != CrossPlatformAddressFamily.AF_INET:
+            continue
+
+        if len(packet.data) < ETHERNET_HEADER_LEN + IPV4_SRC_OFFSET + 4:
+            continue
+
+        src_ip_bytes = packet.data[ETHERNET_HEADER_LEN + IPV4_SRC_OFFSET:ETHERNET_HEADER_LEN + IPV4_SRC_OFFSET + 4]
+        addr = ipaddress.IPv4Address(src_ip_bytes)
+
+        if addr.is_private and not addr.is_loopback and not addr.is_unspecified:
+            print(addr)
+            return
+
+    logger.error("No private IPv4 address found")
+    raise typer.Exit(code=1)
